requirements.txt
langchain==0.1.2
langchain_experimental==0.0.49
langchain-openai==0.0.2
langchain-community
langchain-openai
python-dotenv
streamlit
tabulate
